% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{as.character.tiny_labelled}
\alias{as.character.tiny_labelled}
\alias{as.logical.tiny_labelled}
\alias{as.integer.tiny_labelled}
\alias{as.double.tiny_labelled}
\alias{as.complex.tiny_labelled}
\title{Conversion of Labelled Vectors}
\usage{
\method{as.character}{tiny_labelled}(x, keep_label = TRUE, ...)

\method{as.logical}{tiny_labelled}(x, keep_label = TRUE, ...)

\method{as.integer}{tiny_labelled}(x, keep_label = TRUE, ...)

\method{as.double}{tiny_labelled}(x, keep_label = TRUE, ...)

\method{as.complex}{tiny_labelled}(x, keep_label = TRUE, ...)
}
\arguments{
\item{x}{Object to be coerced}

\item{keep_label}{Logical indicating whether the variable labels and class \code{tiny_labelled} should be kept.}

\item{...}{Further arguments passed to methods}
}
\description{
Functions to convert labelled vectors to other representations.
}
