% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{label_variable}
\alias{label_variable}
\alias{label_variables}
\title{Label Variables Using Pipes}
\usage{
label_variable(x, ...)

label_variables(x, ...)
}
\arguments{
\item{x}{Either a vector or a \code{data.frame}.}

\item{...}{Variable label(s) to be assigned. For data frames, these have to be name-value pairs,
see example.}
}
\description{
\code{label_variable()} can be used to assign variable labels within a workflow
using the tidyverse's pipe operator.
}
\examples{
  library(dplyr)
  test <- npk \%>\%
    label_variable(N = "Nitrogen", P = "Phosphate")
  variable_label(test)
}
