% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabel.R
\name{unlabel}
\alias{unlabel}
\title{Remove Labels from Objects}
\usage{
unlabel(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
Object as \code{x} but without variable labels and with class \code{tiny_labelled} removed.
}
\description{
Remove \link{variable_labels} from a labelled vector or from the columns of a data frame.
}
