% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinylabels.R
\docType{package}
\name{tinylabels}
\alias{tinylabels}
\title{Tiny Variable Labels}
\description{
To learn more about tinylabels, take a look at the vignette: browseVignettes(package = "tinylabels")
}
\section{Maintainer}{

Marius Barth (marius.barth.uni.koeln at gmail.com).
}

